improve compatibility with Vivado synthesis
This commit moves the imports of the `riscv_defines_apu` package from
outside of the module declarations to the inside. This way, the package is
imported only to the modules and not to the entire compilation unit. For
Vivado synthesis, all SystemVerilog files belong to the same compilation
unit, which -- without this commit -- causes errors due to repeated
parameter definitions.